Вместо uint/ulong теперь long

+++ Services/TreeItemComponents/ITreeItemComponents.cs
@@ -4,6 +4,6 @@ namespace PaxDeiBot.Services.TreeItemComponents;
 
 public interface ITreeItemComponents
 {
-    public Task<TreeItemViewModel> GetTree(Guid id, uint multiplier);
-    public Task<List<PrimalComponentViewModel>> GetPrimalComponents(Guid id, uint multiplier);
+    public Task<TreeItemViewModel> GetTree(Guid id, long multiplier);
+    public Task<List<PrimalComponentViewModel>> GetPrimalComponents(Guid id, long multiplier);
 }
